// frontend/src/core/api/generated/model/attendanceEntryViewReason.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */

/**
 * @nullable
 */
export type AttendanceEntryViewReason =
  (typeof AttendanceEntryViewReason)[keyof typeof AttendanceEntryViewReason] | null

export const AttendanceEntryViewReason = {
  SICK: 'SICK',
  UNEXCUSED: 'UNEXCUSED',
  OTHER: 'OTHER',
} as const