// frontend/src/core/api/generated/model/attendanceEntryViewStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type AttendanceEntryViewStatus =
  (typeof AttendanceEntryViewStatus)[keyof typeof AttendanceEntryViewStatus]

export const AttendanceEntryViewStatus = {
  PRESENT: 'PRESENT',
  ABSENT: 'ABSENT',
  EXCUSED: 'EXCUSED',
} as const